backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==0.19.0
requests==2.31.0
gitpython==3.1.40
aiohttp==3.9.1
pydantic==2.5.2
python-multipart==0.0.6
python-magic==0.4.27
